Implemented state purge Light test refactoring Fixed checkpoint deadlocks
