+++ test/DurableTask.SqlServer.Tests/Integration/Orchestrations.cs
@@ -2,60 +2,27 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.Diagnostics;
     using System.Linq;
-    using System.Reflection;
     using System.Threading.Tasks;
     using DurableTask.Core;
     using DurableTask.SqlServer.Tests.Logging;
-    using Microsoft.Extensions.Logging;
-    using Newtonsoft.Json;
+    using DurableTask.SqlServer.Tests.Utils;
     using Newtonsoft.Json.Linq;
     using Xunit;
     using Xunit.Abstractions;
 
     public class Orchestrations : IAsyncLifetime
     {
-        readonly SqlProviderOptions options;
-        readonly TestLogProvider logProvider;
-        readonly ILoggerFactory loggerFactory;
-
-        TaskHubWorker worker;
-        TaskHubClient client;
+        readonly TestService testService;
 
         public Orchestrations(ITestOutputHelper output)
         {
-            Type type = output.GetType();
-            FieldInfo testMember = type.GetField("test", BindingFlags.Instance | BindingFlags.NonPublic);
-            var test = (ITest)testMember.GetValue(output);
-            
-            this.logProvider = new TestLogProvider(output);
-            this.loggerFactory = LoggerFactory.Create(builder =>
-            {
-                builder.AddProvider(this.logProvider);
-            });
-
-            this.options = new SqlProviderOptions
-            {
-                AppName = test.DisplayName,
-                LoggerFactory = this.loggerFactory,
-            };
+            this.testService = new TestService(output);
         }
 
-        async Task IAsyncLifetime.InitializeAsync()
-        {
-            var provider = new SqlOrchestrationService(this.options);
-            await ((IOrchestrationService)provider).CreateIfNotExistsAsync();
+        Task IAsyncLifetime.InitializeAsync() => this.testService.InitializeAsync();
 
-            this.worker = await new TaskHubWorker(provider, this.loggerFactory).StartAsync();
-            this.client = new TaskHubClient(provider, loggerFactory: this.loggerFactory);
-        }
-
-        async Task IAsyncLifetime.DisposeAsync()
-        {
-            await this.worker.StopAsync(isForced: true);
-            this.worker.Dispose();
-        }
+        Task IAsyncLifetime.DisposeAsync() => this.testService.DisposeAsync();
 
         [Fact]
         public async Task EmptyOrchestration_Completes()
@@ -64,7 +31,7 @@
             string orchestrationName = "EmptyOrchestration";
 
             // Does nothing except return the original input
-            StartedInstance<string> instance = await this.RunOrchestration(
+            TestInstance<string> instance = await this.testService.RunOrchestration(
                 input,
                 orchestrationName,
                 implementation: (ctx, input) => Task.FromResult(input));
@@ -73,9 +40,9 @@
                 expectedOutput: input);
 
             // Validate logs
-            LogAssert.NoWarningsOrErrors(this.logProvider);
+            LogAssert.NoWarningsOrErrors(this.testService.LogProvider);
             LogAssert.Sequence(
-                this.logProvider,
+                this.testService.LogProvider,
                 LogAssert.AcquiredAppLock(),
                 LogAssert.CheckpointingOrchestration(orchestrationName));
         }
@@ -88,7 +55,7 @@
             TimeSpan delay = TimeSpan.FromSeconds(3);
 
             // Performs a delay and then returns the input
-            StartedInstance<string> instance = await this.RunOrchestration(
+            TestInstance<string> instance = await this.testService.RunOrchestration(
                 input,
                 orchestrationName,
                 implementation: async (ctx, input) =>
@@ -106,9 +73,9 @@
             Assert.True(state.CreatedTime.Add(delay) <= state.CompletedTime);
 
             // Validate logs
-            LogAssert.NoWarningsOrErrors(this.logProvider);
+            LogAssert.NoWarningsOrErrors(this.testService.LogProvider);
             LogAssert.Sequence(
-                this.logProvider,
+                this.testService.LogProvider,
                 LogAssert.AcquiredAppLock(),
                 LogAssert.CheckpointingOrchestration(orchestrationName),
                 LogAssert.CheckpointingOrchestration(orchestrationName));
@@ -117,7 +84,7 @@
         [Fact]
         public async Task Orchestration_IsReplaying_Works()
         {
-            StartedInstance<string> instance = await this.RunOrchestration<List<bool>, string>(
+            TestInstance<string> instance = await this.testService.RunOrchestration<List<bool>, string>(
                 null,
                 orchestrationName: "TwoTimerReplayTester",
                 implementation: async (ctx, _) =>
@@ -144,12 +111,12 @@
         {
             string input = $"[{DateTime.UtcNow:o}]";
 
-            StartedInstance<string> instance = await this.RunOrchestration(
+            TestInstance<string> instance = await this.testService.RunOrchestration(
                 input,
                 orchestrationName: "OrchestrationWithActivity",
                 implementation: (ctx, input) => ctx.ScheduleTask<string>("SayHello", "", input),
                 activities: new[] {
-                    ("SayHello", MakeActivity((TaskContext ctx, string input) => $"Hello, {input}!")),
+                    ("SayHello", TestService.MakeActivity((TaskContext ctx, string input) => $"Hello, {input}!")),
                 });
 
             OrchestrationState state = await instance.WaitForCompletion(
@@ -162,7 +129,7 @@
         [InlineData(100)]
         public async Task OrchestrationsWithActivityChain_Completes(int parallelCount)
         {
-            List<StartedInstance<string>> instances = await this.RunOrchestrations<int, string>(
+            List<TestInstance<string>> instances = await this.testService.RunOrchestrations<int, string>(
                 parallelCount,
                 _ => null,
                 orchestrationName: "OrchestrationsWithActivityChain",
@@ -177,7 +144,7 @@
                     return value;
                 },
                 activities: new[] {
-                    ("PlusOne", MakeActivity((TaskContext ctx, int input) => input + 1)),
+                    ("PlusOne", TestService.MakeActivity((TaskContext ctx, int input) => input + 1)),
                 });
 
             IEnumerable<Task> tasks = instances.Select(
@@ -193,7 +160,7 @@
             string errorMessage = "Kah-BOOOOOM!!!";
 
             // The exception is expected to fail the orchestration execution
-            StartedInstance<string> instance = await this.RunOrchestration<string, string>(
+            TestInstance<string> instance = await this.testService.RunOrchestration<string, string>(
                 null,
                 orchestrationName: "OrchestrationWithException",
                 implementation: (ctx, input) => throw new Exception(errorMessage));
@@ -208,12 +175,12 @@
         public async Task OrchestrationWithActivityFailure_Fails()
         {
             // Performs a delay and then returns the input
-            StartedInstance<string> instance = await this.RunOrchestration(
+            TestInstance<string> instance = await this.testService.RunOrchestration(
                 null as string,
                 orchestrationName: "OrchestrationWithActivityFailure",
                 implementation: (ctx, input) => ctx.ScheduleTask<string>("Throw", ""),
                 activities: new[] {
-                    ("Throw", MakeActivity<string, string>((ctx, input) => throw new Exception("Kah-BOOOOOM!!!"))),
+                    ("Throw", TestService.MakeActivity<string, string>((ctx, input) => throw new Exception("Kah-BOOOOOM!!!"))),
                 });
 
             OrchestrationState state = await instance.WaitForCompletion(
@@ -224,7 +191,7 @@
         [Fact]
         public async Task OrchestrationWithActivityFanOut()
         {
-            StartedInstance<string> instance = await this.RunOrchestration<string[], string>(
+            TestInstance<string> instance = await this.testService.RunOrchestration<string[], string>(
                 null,
                 orchestrationName: "OrchestrationWithActivityFanOut",
                 implementation: async (ctx, _) =>
@@ -241,7 +208,7 @@
                     return results;
                 },
                 activities: new[] {
-                    ("ToString", MakeActivity((TaskContext ctx, int input) => input.ToString())),
+                    ("ToString", TestService.MakeActivity((TaskContext ctx, int input) => input.ToString())),
                 });
 
             OrchestrationState state = await instance.WaitForCompletion(
@@ -255,7 +222,7 @@
         {
             TaskCompletionSource<int> tcs = null;
 
-            StartedInstance<string> instance = await this.RunOrchestration<int, string>(
+            TestInstance<string> instance = await this.testService.RunOrchestration<int, string>(
                 null,
                 orchestrationName: "OrchestrationWithExternalEvents",
                 implementation: async (ctx, _) =>
@@ -295,7 +262,7 @@
             TimeSpan delay = TimeSpan.FromSeconds(30);
 
             // Performs a delay and then returns the input
-            StartedInstance<string> instance = await this.RunOrchestration(
+            TestInstance<string> instance = await this.testService.RunOrchestration(
                 input,
                 orchestrationName,
                 implementation: (ctx, input) => ctx.CreateTimer(ctx.CurrentUtcDateTime.Add(delay), input));
@@ -313,9 +280,9 @@
                 expectedOutput: "Bye!");
 
             // Validate logs
-            LogAssert.NoWarningsOrErrors(this.logProvider);
+            LogAssert.NoWarningsOrErrors(this.testService.LogProvider);
             LogAssert.Sequence(
-                this.logProvider,
+                this.testService.LogProvider,
                 LogAssert.AcquiredAppLock(),
                 LogAssert.CheckpointingOrchestration(orchestrationName),
                 LogAssert.CheckpointingOrchestration(orchestrationName));
@@ -325,7 +292,7 @@
         [Fact]
         public async Task ContinueAsNew()
         {
-            StartedInstance<int> instance = await this.RunOrchestration(
+            TestInstance<int> instance = await this.testService.RunOrchestration(
                 input: 0,
                 orchestrationName: "ContinueAsNewTest",
                 implementation: async (ctx, input) =>
@@ -341,260 +308,6 @@
 
             TimeSpan timeout = TimeSpan.FromSeconds(5);
             await instance.WaitForCompletion(timeout, expectedOutput: 10, continuedAsNew: true);
-        }
-
-        async Task<List<StartedInstance<TInput>>> RunOrchestrations<TOutput, TInput>(
-            int count,
-            Func<int, TInput> inputGenerator,
-            string orchestrationName,
-            Func<OrchestrationContext, TInput, Task<TOutput>> implementation,
-            Action<OrchestrationContext, string, string> onEvent = null,
-            params (string name, TaskActivity activity)[] activities)
-        {
-            // Register the inline orchestration - note that this will only work once per test
-            Type orchestrationType = typeof(OrchestrationShim<TOutput, TInput>);
-
-            this.worker.AddTaskOrchestrations(new TestObjectCreator<TaskOrchestration>(
-                orchestrationName,
-                MakeOrchestration(implementation, onEvent)));
-
-            foreach ((string name, TaskActivity activity) in activities)
-            {
-                this.worker.AddTaskActivities(new TestObjectCreator<TaskActivity>(name, activity));
-            }
-
-            DateTime utcNow = DateTime.UtcNow;
-
-            var instances = new List<StartedInstance<TInput>>(count);
-            for (int i = 0; i < count; i++)
-            {
-                TInput input = inputGenerator(i);
-                OrchestrationInstance instance = await this.client.CreateOrchestrationInstanceAsync(
-                    orchestrationName,
-                    string.Empty /* version */,
-                    input);
-
-                // Verify that the CreateOrchestrationInstanceAsync implementation set the InstanceID and ExecutionID fields
-                Assert.NotNull(instance.InstanceId);
-                Assert.NotNull(instance.ExecutionId);
-
-                instances.Add(new StartedInstance<TInput>(this.client, instance, utcNow, input));
-            }
-            
-            return instances;
-        }
-
-        async Task<StartedInstance<TInput>> RunOrchestration<TOutput, TInput>(
-            TInput input,
-            string orchestrationName,
-            Func<OrchestrationContext, TInput, Task<TOutput>> implementation,
-            Action<OrchestrationContext, string, string> onEvent = null,
-            params (string name, TaskActivity activity)[] activities)
-        {
-            var instances = await this.RunOrchestrations(
-                count: 1,
-                inputGenerator: i => input,
-                orchestrationName,
-                implementation,
-                onEvent,
-                activities);
-
-            return instances.First();
-        }
-
-        static TaskOrchestration MakeOrchestration<TOutput, TInput>(
-            Func<OrchestrationContext, TInput, Task<TOutput>> implementation,
-            Action<OrchestrationContext, string, string> onEvent = null)
-        {
-            return new OrchestrationShim<TOutput, TInput>(implementation, onEvent);
-        }
-
-        // This is just a wrapper around the constructor for convenience. It allows us to write 
-        // less code because generic arguments for methods can be implied, unlike constructors.
-        static TaskActivity MakeActivity<TInput, TOutput>(
-            Func<TaskContext, TInput, TOutput> implementation)
-        {
-            return new ActivityShim<TInput, TOutput>(implementation);
-        }
-
-        static string GetFriendlyTypeName(Type type)
-        {
-            string friendlyName = type.Name;
-            if (type.IsGenericType)
-            {
-                int iBacktick = friendlyName.IndexOf('`');
-                if (iBacktick > 0)
-                {
-                    friendlyName = friendlyName.Remove(iBacktick);
-                }
-
-                friendlyName += "<";
-                Type[] typeParameters = type.GetGenericArguments();
-                for (int i = 0; i < typeParameters.Length; ++i)
-                {
-                    string typeParamName = GetFriendlyTypeName(typeParameters[i]);
-                    friendlyName += (i == 0 ? typeParamName : "," + typeParamName);
-                }
-
-                friendlyName += ">";
-            }
-
-            return friendlyName;
-        }
-
-        class ActivityShim<TInput, TOutput> : TaskActivity<TInput, TOutput>
-        {
-            public ActivityShim(Func<TaskContext, TInput, TOutput> implementation)
-            {
-                this.Implementation = implementation;
-            }
-
-            public Func<TaskContext, TInput, TOutput> Implementation { get; }
-
-            protected override TOutput Execute(TaskContext context, TInput input)
-            {
-                return this.Implementation(context, input);
-            }
-        }
-
-        class OrchestrationShim<TOutput, TInput> : TaskOrchestration<TOutput, TInput>
-        {
-            public OrchestrationShim(
-                Func<OrchestrationContext, TInput, Task<TOutput>> implementation,
-                Action<OrchestrationContext, string, string> onEvent = null)
-            {
-                this.Implementation = implementation;
-                this.OnEventRaised = onEvent;
-            }
-
-            public Func<OrchestrationContext, TInput, Task<TOutput>> Implementation { get; set; }
-
-            public Action<OrchestrationContext, string, string> OnEventRaised { get; set; }
-
-            public override Task<TOutput> RunTask(OrchestrationContext context, TInput input)
-                => this.Implementation(context, input);
-
-            public override void RaiseEvent(OrchestrationContext context, string name, string input)
-                => this.OnEventRaised(context, name, input);
-        }
-
-        class StartedInstance<T>
-        {
-            readonly TaskHubClient client;
-            readonly OrchestrationInstance instance;
-            readonly DateTime startTime;
-            readonly T input;
-
-            public StartedInstance(
-                TaskHubClient client,
-                OrchestrationInstance instance,
-                DateTime startTime,
-                T input)
-            {
-                this.client = client;
-                this.instance = instance;
-                this.startTime = startTime;
-                this.input = input;
-            }
-
-            OrchestrationInstance GetInstanceForAnyExecution() => new OrchestrationInstance
-            {
-                InstanceId = this.instance.InstanceId,
-            };
-
-            public async Task<OrchestrationState> WaitForCompletion(
-                TimeSpan timeout = default,
-                OrchestrationStatus expectedStatus = OrchestrationStatus.Completed,
-                object expectedOutput = null,
-                string expectedOutputRegex = null,
-                bool continuedAsNew = false)
-            {
-                if (timeout == default)
-                {
-                    timeout = TimeSpan.FromSeconds(5);
-                }
-
-                if (Debugger.IsAttached)
-                {
-                    timeout = timeout.Add(TimeSpan.FromMinutes(5));
-                }
-
-                OrchestrationState state = await this.client.WaitForOrchestrationAsync(this.GetInstanceForAnyExecution(), timeout);
-                Assert.NotNull(state);
-                Assert.Equal(expectedStatus, state.OrchestrationStatus);
-
-                if (this.input != null)
-                {
-                    Assert.Equal(JToken.FromObject(this.input), JToken.Parse(state.Input));
-                }
-                else
-                {
-                    Assert.Null(state.Input);
-                }
-
-                // For created time, account for potential clock skew
-                Assert.True(state.CreatedTime >= this.startTime.AddMinutes(-5));
-                Assert.True(state.LastUpdatedTime > state.CreatedTime);
-                Assert.True(state.CompletedTime > state.CreatedTime);
-                Assert.NotNull(state.OrchestrationInstance);
-                Assert.Equal(this.instance.InstanceId, state.OrchestrationInstance.InstanceId);
-
-                if (continuedAsNew)
-                {
-                    Assert.NotEqual(this.instance.ExecutionId, state.OrchestrationInstance.ExecutionId);
-                }
-                else
-                {
-                    Assert.Equal(this.instance.ExecutionId, state.OrchestrationInstance.ExecutionId);
-                }
-
-                if (expectedOutput != null)
-                {
-                    Assert.NotNull(state.Output);
-                    try
-                    {
-                        // DTFx usually encodes outputs as JSON values. The exception is error messages.
-                        // If this is an error message, we'll throw here and try the logic in the catch block.
-                        JToken.Parse(state.Output);
-                        Assert.Equal(JToken.FromObject(expectedOutput).ToString(Formatting.None), state.Output);
-                    }
-                    catch (JsonReaderException)
-                    {
-                        Assert.Equal(expectedOutput, state?.Output);
-                    }
-                }
-
-                if (expectedOutputRegex != null)
-                {
-                    Assert.Matches(expectedOutputRegex, state.Output);
-                }
-
-                return state;
-            }
-
-            internal Task RaiseEventAsync(string name, object value)
-            {
-                return this.client.RaiseEventAsync(this.instance, name, value);
-            }
-
-            internal Task TerminateAsync(string reason)
-            {
-                return this.client.TerminateInstanceAsync(this.instance, reason);
-            }
-        }
-
-        class TestObjectCreator<T> : ObjectCreator<T>
-        {
-            readonly T obj;
-
-            public TestObjectCreator(string name, T obj)
-            {
-                this.Name = name;
-                this.Version = string.Empty;
-                this.obj = obj;
-            }
-
-            public override T Create() => this.obj;
         }
     }
 }
